refactor: move helper to separate file
